Change how require_once is adding libraries
There is a conflict with event manager and SOLR. I raised an issue and Jeroen Dalsem suggested this change. There is a conflict in how Elgg handles require_once with files with the exact name. I tested this and that solves the conflict. Other plugins also use this structure and would conflict with SOLR too